refactor: promote the built-in dataset into its own package
Move _builtin_data and _flop_weights_tree_view into a new counting/builtin_data/ package (as _dataset and _tree_view), so config imports them from a sibling package instead of reaching up to a module sitting beside it -- keeping the dependency pointing downward through the folder tree.
