+++ pyproject.toml
@@ -284,7 +284,7 @@ source_paths = ["src/counted_float"]
 # Excluded from mutation -- every surviving mutant here is a false signal, not a real gap:
 #   - _callsite.py is pure call-stack introspection, and mutmut runs tests through a trampoline that
 #     becomes the reported call site, so no test can exercise it.
-#   - _flop_weights_tree_view.py is a renderer whose meaningful behavior is covered elsewhere: the
+#   - builtin_data/_tree_view.py is a renderer whose meaningful behavior is covered elsewhere: the
 #     plain-text layout byte-for-byte by the docs-drift test (which mutmut can't run -- it regenerates
 #     via a subprocess), and the color styling + integer formatting (which the drift test strips /
 #     never exercises) by tests/counting/test_flop_weights_tree_view.py. What remains is purely
@@ -301,7 +301,7 @@ source_paths = ["src/counted_float"]
 do_not_mutate = [
     "src/counted_float/_core/counting/verbosity/_callsite.py",
     "src/counted_float/_core/counting/verbosity/_output.py",
-    "src/counted_float/_core/counting/_flop_weights_tree_view.py",
+    "src/counted_float/_core/counting/builtin_data/_tree_view.py",
     "src/counted_float/_core/benchmarking/flops/_flops_probes.py",
 ]
 # -n 0 runs the suite serially (mutmut parallelizes across mutants itself); it overrides the
