+++ src/counted_float/_core/counting/__init__.py
@@ -1,4 +1,4 @@
-from ._builtin_data import BuiltInData
 from ._context_managers import FlopCountingContext, PauseFlopCounting
 from ._counted_float import CountedFloat
+from .builtin_data import BuiltInData
 from .verbosity import Verbosity
